refactor: move expr translations under expressions folder
