+++ samples/demo/index.queries.ts
@@ -1,0 +1,15 @@
+
+
+export type SomeInputQueryParams = [];
+
+
+export interface ISomeInputQueryResult {
+    
+};
+
+
+export interface ISomeInputQueryQuery {
+    params: SomeInputQueryParams;
+    result: ISomeInputQueryResult;
+};
+
