+++ samples/demo/index.ts
@@ -1,5 +1,12 @@
 import { sql } from "sqlx-ts";
 
+const someInputQuery = sql`
+INSERT INTO items (id, food_type, time_takes_to_cook, table_id, points)
+VALUES
+(?, 'australian', 2, 1, 2),
+(?, 'test', 2, 1, 2);
+`
+
 /*
 const someQuery = sql`
 SELECT *
@@ -38,7 +45,7 @@ FROM items
 `
 */
 
-
+/*
 // Querying from an unknown table
 const someQuery = sql`SELECT * FROM indexjs_unknown`;
 
@@ -54,3 +61,5 @@ VALUES ('steak', 1, 1, 1, 1);
 if (true) {
 const query3 = sql`SELECT * FROM if_statement1;`;
 }
+*/
+
