--- samples/parameterized-query/index.queries.ts
@@ -1,47 +1,0 @@
-
-
-export type SomeQueryParams = [];
-
-
-export interface ISomeQueryResult {
-    
-};
-
-
-export interface ISomeQueryQuery {
-    params: SomeQueryParams;
-    result: ISomeQueryResult;
-};
-
-
-
-
-export type InsertQueryParams = [];
-
-
-export interface IInsertQueryResult {
-    
-};
-
-
-export interface IInsertQueryQuery {
-    params: InsertQueryParams;
-    result: IInsertQueryResult;
-};
-
-
-
-
-export type Query3Params = [];
-
-
-export interface IQuery3Result {
-    
-};
-
-
-export interface IQuery3Query {
-    params: Query3Params;
-    result: IQuery3Result;
-};
-
