+++ src/ts_generator/sql_parser/expressions/translate_table_with_joins.rs
