+++ src/ts_generator/sql_parser/expressions/translate_wildcard_expr.rs
