+++ src/ts_generator/sql_parser/translate_insert.rs
@@ -1,0 +1,1 @@
+
